+++ apps/docs/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @threlte/docs-next
 
+## 0.6.0
+
+### Minor Changes
+
+- 144b55a: Deprecated the component `<Center>` as it's being renamed to `<Align>`. Renamed the prop `autoCenter` to `auto`, renamed events and slot props accordingly.
+
 ## 0.5.0
 
 ### Minor Changes